% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLoops.R
\name{loadHandClusters}
\alias{loadHandClusters}
\title{Load pre-defined clusters}
\usage{
loadHandClusters()
}
\value{
A list of grouped clusters
}
\description{
Load a pre-specified cluster list that can be called by handMetaCluster()
}
\examples{
handCluster<-loadHandClusters()  
}
