% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLoops.R
\name{scMatrixTest}
\alias{scMatrixTest}
\title{Generate all the signature matrices one time}
\usage{
scMatrixTest(
  exprData,
  randomize = TRUE,
  skipShrink = FALSE,
  proportional = FALSE,
  handMetaCluster = NULL
)
}
\arguments{
\item{exprData}{The gene express data. Each row is a gene, and each column is an example of a particular cell type.}

\item{randomize}{Set to TRUE randomize the sets selected by ADAPTS::scSample (DEFAULT: TRUE)}

\item{skipShrink}{Set to TRUE to skip shrinking the signatrure matrix (DEFAULT: TRUE)}

\item{proportional}{Set to true to make the training set cell type proportional.  Ignores group size (DEFAULT: FALSE)}

\item{handMetaCluster}{Load in pre-defined meta clusters. Set to NULL to automatically group indistinguishable 
cells into same cluster use clustWspillOver(DEFAULT: NULL)}
}
\value{
A list of results including prediction accuracy and cell enrichment
}
\description{
This wrapper is helpful for repetively matrix generation. It generates seed matrix, all-gene matrix, augmneted matrix, shrunk matrix,
and all the clustered matrices in one call.
}
