% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLoops.R
\name{loopTillConvergence}
\alias{loopTillConvergence}
\title{Loop scMatrixtest until convergence}
\usage{
loopTillConvergence(numLoops, fastStop, exprData, changePer, handMetaCluster)
}
\arguments{
\item{numLoops}{The number of iterations. Set to null to loop until results converge.}

\item{fastStop}{Set to TRUE to break the loop when correlation, correlation spear, mae and rmse all converge.}

\item{exprData}{The single cell matrix}

\item{changePer}{The maximum percentage of change allowed for convergence#'}

\item{handMetaCluster}{Load in pre-defined meta clusters. Set to NULL to automatically group indistinguishable 
cells into same cluster use clustWspillOver(DEFAULT: NULL)}
}
\value{
A list of results generated from all the iterative calls of scMatrixTest
}
\description{
Iteratively call scMatrixtest numLoops times with the option to fast stop 
if correlation, correlation spear, mae and rmse all converge
}
